requirements.txt
absl-py==2.3.1
anyio==4.12.1
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
grpcio==1.75.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
ImageIO==2.37.2
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
numpy==2.2.6
openai-clip==1.0.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==4.21.2
psutil==7.1.1
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
